+++ src/data_strcture/sort/quick_sort.py
@@ -16,7 +16,7 @@ def quick_sort(input_list):
 
 def main():
     u_list = input("input a list of init\n")
-    u_list = list(map(int, u_list.split(",")))
+    u_list = list(map(float, u_list.split(",")))
     print(quick_sort(u_list))
